refac: 全てをリザルトクラスにいれなくてもいいようにした
